fix: address review feedback for wc/v3 orders isolation tests
- Skip created_via filter tests on WC < 9.6 (parameter not supported)
- Extract shared helpers into WC_REST_Order_Helpers trait
- Use per_page instead of posts_per_page for REST API requests
- Add docstrings to all test methods and classes
